Additional template partials and helpers extraction (#8205)
* Add action based authorized helper methods

These are common enough for helper methods. We'll also be adding some to a new partial so this will clean things up a bit. Although these are public methods that users can use as documented so its not really hiding anything, more for convenience.

* Extract index-as-table default actions to partial

This way users can customize this e.g. icons instead of text. These now use new helper methods to keep this clean.

This also removes the title attribute since its advised not to include it if its the same text as the link label since its verbose/redundant (screen readers).

Insert TableActions without builder_method since builder_method is just a shorthand around insert_tag, we can use that instead which is best otherwise we expose `table_actions` globally meaning any user can call it in their code but shouldn't be able too. It also takes up a name in Arbre. We noticed this with the original header/footer components for the site, where it would clash when trying to use header/footer HTML elements through Arbre.

* Use new action authorized methods elsewhere

* Rename has-many CSS classes

* Update Tailwind content paths

By scanning the plugin.js file, we won't need to scan the `/lib/active_admin/*` path since our component classes are already listed in the plugin as Tailwind finds each correctly and adds to CSS. If it becomes an issue, we can add the original path back.

For the host app, add specific paths for active_admin and admin (default namespace) rather than scan entire app/views directory.

Remove scanning CSS files as Tailwind content configuration docs state to never do this.

> It’s also important that you don’t scan any CSS files — configure Tailwind to scan your templates where your class names are being used, never the CSS file that Tailwind is generating.

https://tailwindcss.com/docs/content-configuration#pattern-recommendations

* Rename variable to activeAdminPath for clarity

Should help on the app host side since "activeAdmin" in the name to indicate that its scanning the files in the gem. Since there are complimentary ActiveAdmin gems, for sure those will update and need to be included in this file in the host app so they can follow a convention/pattern.

* Use attributes_table_for throughout

We'll update the docs separately from this changeset.

* Rename to index_helper

We'll be adding new helper methods that are used on the index view like scope_name is (scopes and active filters components).

* Extract batch actions dropdown to partial

With this, we remove the only use of the private DropdownMenu component which can now be removed.

* Remove dropdown_menu component

* Rename CSS class as batch-actions-dropdown

With dropdown_menu removed, rename this class so its specific to batch actions.

* Remove flag from sample resource

* Fix test coverage by using item DSL

This shows some of that bad DSLs we have since we only tested this using link_to but it doesn't work with multiple links here, you have to use the item method the actions method (table_actions component) provides.

* Migrate collection helpers to app/helpers

Some uses already rely on the `collection` helper method (InheritedResources::Base) as a default argument but others through a provided argument which happens to be the same name (!). We only use these 2 methods in views so we can define as a plain Rails helper.

Since we have a feature spec using the paginated_collection in a page, we have to include the helper in our base controller so it's available to both resources and pages.

* Extract active filters to partial

This will allow us to update this further with autoloading. This also moves the scopes to the model class to keep together. No need for accessors so changed to attribute readers.

* Remove builder_method from internal components

These components are not meant to be used by users directly, only internally but by declaring a builder_method we take up that name in a global space that would compete with what users want to create. We had that originally with columns/column and header/footer components https://github.com/activeadmin/activeadmin/discussions/8135 we will create these internal only components using insert_tag and provide the class leaving builder methods open to users and third party libraries.