Update data-table-sorted-icon styles
Trying this out in a few apps wasn't always picked up. It doesn't follow how we typically do other CSS here so this should play nicer with Tailwind's content look up so no need to specify these in a safelist. Since this is already prefixed with data-table- in the name, no need to scope to the data-table class, like we do with data-table-resource-actions and other classes.

This also sets a fixed width and height on the icon since in some cases the SVG can get cut off at the bottom portion. It's not replicable on our dev generated app, or on demo but can occur.