Consolidate bug report template workflow
The script now runs quickly since we no longer have to install sassc. This used to be much slower. For example, runs for a couple of months ago, could be as slow as 3 minutes whereas now it runs in about 30 seconds or less.

GitHub no longer documents the approach we took with the empty workflow and paths/paths-ignore so we'll consolidate into one workflow using the changes-files action.

Since this is now fast, we want to run this on any PR with applicable files changed to get feedback, rather than fail on a schedule and realize then that something is wrong. In this case it was the release of importmap-rails v2 and a deleted argument.