Template and styles fixes (#8207)
* Remove shim arg - deleted in importmap-rails v2

This argument has been removed so we can't default to it. For now we intend to support v1 and v2 of importmap-rails since breaking changes are only around defaults. If users don't want the shim but want to remain on v1 they can modify this template.

* Lock bug report template to use importmap-rails v2

The previous change fixes the bug report template workflow failure but here we'll want to stick to the current version.

* Update data-table-sorted-icon styles

Trying this out in a few apps wasn't always picked up. It doesn't follow how we typically do other CSS here so this should play nicer with Tailwind's content look up so no need to specify these in a safelist. Since this is already prefixed with data-table- in the name, no need to scope to the data-table class, like we do with data-table-resource-actions and other classes.

This also sets a fixed width and height on the icon since in some cases the SVG can get cut off at the bottom portion. It's not replicable on our dev generated app, or on demo but can occur.

* Consolidate bug report template workflow

The script now runs quickly since we no longer have to install sassc. This used to be much slower. For example, runs for a couple of months ago, could be as slow as 3 minutes whereas now it runs in about 30 seconds or less.

GitHub no longer documents the approach we took with the empty workflow and paths/paths-ignore so we'll consolidate into one workflow using the changes-files action.

Since this is now fast, we want to run this on any PR with applicable files changed to get feedback, rather than fail on a schedule and realize then that something is wrong. In this case it was the release of importmap-rails v2 and a deleted argument.