Use library specific tailwind config file
We want to generate this with its own name and not assume the host app isn't using Tailwind already. If they aren't and only plan to for the admin, then its a simple change on their to rename this.

We don't need to include Procfile.dev or builds/.keep since the expectation is that the host app has installed and run the setup for both cssbundling-rails (tailwind) and importmap-rails.

The most a host app will have to do after running the assets generator, is just add update the build:css script in package.json with the ActiveAdmin specific Tailwind build.