Assets generator update (#8208)
* Remove css flag as it uses tailwindcss-rails gem

There doesn't seem to be a setting for cssbundling-rails with tailwind as this always uses tailwindcss-rails and seems by design. Since we want to install and setup importmap-rails anyway, we'll do the same with cssbundling-rails.

* Add cssbundling-rails and run setup

* Run importmap-rails setup

We don't need to install the gem since we pass the --javascript=importmap argument on generating the dev/test app already.

* Use library specific tailwind config file

We want to generate this with its own name and not assume the host app isn't using Tailwind already. If they aren't and only plan to for the admin, then its a simple change on their to rename this.

We don't need to include Procfile.dev or builds/.keep since the expectation is that the host app has installed and run the setup for both cssbundling-rails (tailwind) and importmap-rails.

The most a host app will have to do after running the assets generator, is just add update the build:css script in package.json with the ActiveAdmin specific Tailwind build.

* Update Gemfile.lock

This now happens when running the install/setup scripts for cssbundling-rails and importmap-rails so we'll include this update.

* Remove package.json template from assets gen

This is no longer needed as all we need to do and for users, is to add the activeadmin npm package and update the build:css command with the updated tailwind command for activeadmin CSS. For that we the `npm pkg set scripts` command.

* Remove old skip flags from rails generate

These are no longer defined. See if removing the bundle/gemfile resolves the Rails 6.1 CI build failure as everything else is passing.

* Add --skip-git flag to rails new

We don't need git here since the app is created within the tmp directory and already ignored. It also would be one less action to run and less output too.

* Simplify with environment/route methods

No need to do these updates manually as Rails already includes template methods for these actions.

https://guides.rubyonrails.org/rails_application_templates.html#environment-application-data-nil-options-block

https://guides.rubyonrails.org/rails_application_templates.html#route-routing-code