Create initial Upgrading guide for v4 (#8182)
* Create UPGRADING.md

* Update assets gen to remove old SCSS/JS files

* Add views generator to copy partials

For now this copies the partials that we consider public and is expected that to be customized in the host app unless the defaults are fine as is.

Any other partials are considered private. If they are copied, the user is expected to update those per release as they can encounter breaking changes, much like with monkey patching.