Replace link-default with base <a> style
This rule starts to get problematic with links elsewhere that are set to inherit text color in default state (non-dark mode). I noticed this with the table column links. We've already been setting no-underline on any links elsewhere, including table column. This is easier to override per element and will at least give links a default style similar to the previous version which I think others will still want to have in place.