Fix bug with page title on form submission
When submitting the form and its re-rendered due to validation errors, then the page title ending up being a generic "Create" vs what it used to be like "New Post". Reviewing the original implementation in v3, the action is normalized regardless of the action so we have to account for create/update here in the when statement to preverse the same behavior.