Some final updates before v4 beta1 (#8210)
* Add name presence validation for testing

We should have at least a model with validation for form testing without having to make changes. For now we'll update the category model since it's the easiest. Only one case where it was creating the model without any data.

* Add custom-index class with data attribute

Use an explicit named class name and show adding a data attribute as an example here.

* Rename toggle/selection classes for batch actions

* Replace link-default with base <a> style

This rule starts to get problematic with links elsewhere that are set to inherit text color in default state (non-dark mode). I noticed this with the table column links. We've already been setting no-underline on any links elsewhere, including table column. This is easier to override per element and will at least give links a default style similar to the previous version which I think others will still want to have in place.

* Set text-sm on forms

* Fix bug with page title on form submission

When submitting the form and its re-rendered due to validation errors, then the page title ending up being a generic "Create" vs what it used to be like "New Post". Reviewing the original implementation in v3, the action is normalized regardless of the action so we have to account for create/update here in the when statement to preverse the same behavior.

* Remove sidebar from posts show page

* Remove index_table class with improved tests

Tests for page registration which was reusing the index_table class directly to work with an existing expectation but its using table_for which we have our dedicated class name for since its a table component vs just a table.