Include additional files in gem release
At first we were tackling a sprockets error but it become evident more was missing that should be included.

We could have the tailwind config point to the npm package plugin.js for the content but since its already from the gem, best to just include it. It's how we expected it to work.

We need to include the config directory for our importmap config.

While those were the only two required changes, I figured we should have the UPGRADING guide set too.