Vendor rails-ujs ESM JS (#8217)
With this we avoid any files from CDN and issues with CSP. We intended to do this originally but couldn't get it to work. The docs in importmaps-rails were corrected this past week to show what to do in this case. Previously, I was setting the "to:" with the vendor path but it's unnecessary which is why it didn't work.