Streamline release rake with just npm_push task
For npm push we can just rely on the npm package itself to get the version and this just becomes a few lines. No manager needed.

We'd rather have releasing mimic common steps (e.g. creating a branch, making changes, committing) vs automation that is in rake and has to be specific to each and every version type. We used a beta release which didn't fit in here. Further rake has weird and hard to grasp handling of arguments. It's much easier to use a small shell script that can also use Ruby which is why we have bin/prep-release now.

Further this added a significant amount of code around releasing when it's unnecessary and we don't want to maintain so much.

This also removes chandler usage. We'll draft releases in GitHub Releases manually since it's automated now after selecting the version tag and will auto generate the changes history.