Remove gemspec spec
This has caused several flaky failures now that we are on a beta release. While this is helpful its not necessary nor adds much value in the test suite. The reason for adding this was to not have it report warnings but we had to allow that when we changed the gemspec to have some open ended dependencies. This was done knowing the trade offs because we want users to be able to install activeadmin on newer versions of those dependencies to try it out and contribute a fix rather than opening issues and asking for an update, reducing the burden on maintainers.

As part of the releasing process, we do this as an optional step anyway to check the contents of the release. Also releasing itself has to build the gem which has no issues and hasn't so we'd know if it did.

Introduced in: https://github.com/activeadmin/activeadmin/pull/5634