Simplify release process (#8219)
* Add bin/prep-release script

This reuses the bin/bundle script to handle bundle install on every gemfile after updating the version. No need to edit lockfiles directly. This will replace the release manager rake tasks.

* Streamline release rake with just npm_push task

For npm push we can just rely on the npm package itself to get the version and this just becomes a few lines. No manager needed.

We'd rather have releasing mimic common steps (e.g. creating a branch, making changes, committing) vs automation that is in rake and has to be specific to each and every version type. We used a beta release which didn't fit in here. Further rake has weird and hard to grasp handling of arguments. It's much easier to use a small shell script that can also use Ruby which is why we have bin/prep-release now.

Further this added a significant amount of code around releasing when it's unnecessary and we don't want to maintain so much.

This also removes chandler usage. We'll draft releases in GitHub Releases manually since it's automated now after selecting the version tag and will auto generate the changes history.

* Remove changelog rake task

* Remove release gem group

* Move to GitHub Releases for changelogs

* Remove gemspec spec

This has caused several flaky failures now that we are on a beta release. While this is helpful its not necessary nor adds much value in the test suite. The reason for adding this was to not have it report warnings but we had to allow that when we changed the gemspec to have some open ended dependencies. This was done knowing the trade offs because we want users to be able to install activeadmin on newer versions of those dependencies to try it out and contribute a fix rather than opening issues and asking for an update, reducing the burden on maintainers.

As part of the releasing process, we do this as an optional step anyway to check the contents of the release. Also releasing itself has to build the gem which has no issues and hasn't so we'd know if it did.

Introduced in: https://github.com/activeadmin/activeadmin/pull/5634