Update gherkin-lint workflow to use changed-files
GitHub doesn't document this solution anymore for a workflow that is a required PR check but optimizing it to only run when applicable files are changed. We now use changed-files action to replace this and use a single workflow, so the workflow will always run on PR but key steps may not run if applicable files haven't been changed replicating previous setup.