Update form errors styles
Now that these are visible on submission when form is invalid, we can update them. If the list only contains a single item them the bottom margin makes it look off. We now use the space-y-* utility to dynamically control the spacing between child elements from the parent. Also adjust the bottom margin on the list to be larger and smaller rounded borders to match.