Ensure form presenter is applied on create/update actions (#8238)
* Apply form page presenter on submit when invalid

If the form is invalid, the form page presenter wasn't being applied. Same fix as we did for the page title on form submission in https://github.com/activeadmin/activeadmin/pull/8210 specifically this commit: https://github.com/activeadmin/activeadmin/pull/8210/commits/783992e2a9679c6a3230edfb98a0acdcc2664083

* Update form errors styles

Now that these are visible on submission when form is invalid, we can update them. If the list only contains a single item them the bottom margin makes it look off. We now use the space-y-* utility to dynamically control the spacing between child elements from the parent. Also adjust the bottom margin on the list to be larger and smaller rounded borders to match.