Use an anchor for tabs component toggle (#8242)
Update tabs to use anchors instead of buttons

If using tabs in a form, the button styles will conflict. Considering the trade offs, we'll just go with anchors instead for the tab toggles.

Based on Flowbite's interactive example while it uses buttons, other examples use anchors, so it works fine.