Add `use_static_assets` feature
Related to #8246.

This increases the gem size from 173 to 203 KB (as a side note, it might be worth investigating why it is already so big).

Set `use_static_assets` to true to test. I tested it with sprockets, propshaft, and the default pipeline. Of course this would need some automated tests as well. It could also use a mention in the documentation (UPGRADING.md? web docs?), and possibly a generator flag.