Fix style for multiple select inputs
I had a problem where a <select multiple> element was getting
a white background with white text when in dark mode.

It looks like it was a CSS specificity issue because the rule matches
both the `select` and `[multiple]` rules.

The rule further down is supposed to kick in i.e.
```
      [['[type=date]', '[type=email]', '[type=number]', '[type=password]', '[type=tel]', '[type=text]', '[type=time]', '[type=url]', 'select', 'textarea']]: {
        '@apply bg-gray-50 border border-gray-300 text-gray-900 rounded-md focus:ring-blue-500 focus:border-blue-500 w-full dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500': {}
```

.. but the specificity of the base CSS rule was getting in the way of applying `dark:bg-gray-700`