TableFor, add tbody_data and row_data options
This patch adds a `tbody_html` and `row_data` options to the `TableFor` (and
`IndexAsTable`) classes. These options allow you to pass a hash of html options
and data attributes to the `tbody` and `tr` elements, respectively.

This is useful when you want to add features to the table that require data
attributes. For example, adding a sortable row via JavaScript.

You can see how I'm currently using this approach in [this app](https://github.com/acp-admin/acp-admin):
- [monkey patching](https://github.com/acp-admin/acp-admin/blob/master/config/initializers/active_admin.rb#L261-L328) (same approach as this PR)
- [using both options on a `table_for` method](https://github.com/acp-admin/acp-admin/blob/master/app/views/active_admin/deliveries/_baskets.html.arb#L10-L13) for using [Stimulus Sortable component](https://www.stimulus-components.com/docs/stimulus-sortable/).