Remove parallel_rspec, faster run without
Now that we've removed a lot of code and thus tests with v4, there's less value in using parallel_tests for rspec when the rspec tests take just 10 seconds to run sequentially. Comparing when running rspec in parallel, sequential runs now take 11-12 seconds less.

We'll continue to use parallel_tests for cucumber since those are slower, although in parallel its just 2 seconds faster.

We'll try this on CI to compare.