Skip tests if not making code changes
We could write this a few different ways. Fewer paths to specify if we target the paths we know we don't want to run the tests for. The focus here is on the docs and any Markdown files outside that directory (e.g. README.md or PR template in .github).