Try importing each jquery-ui file dependency
I want to see if this works over including the full exported source since we'd be locked to that version. This way, the user's jquery-ui version would be applied but not sure if it will work in an app so we'll need to test. Isolating change in case we want to revert it.