Use safe navigation for `pundit_default_policy`
Use the safe navigation operator (`&.`) to streamline the
`pundit_default_policy` constantization check, removing the need for a
redundant `&&`.

This change is now possible as the codebase no longer requires Ruby 2.3
support, which did not yet support safe navigation.

The previous implementation was added in commit c93f6f4 to ensure
compatibility with older Ruby versions.

Also add the safe navigation operator to a subsequent usage of
`default_policy_class`