Improve v3 docs regarding compatibility with vite_rails (#8549)
Active Admin v3 can be used with the vite_rails gem, but the
instructions for doing so are buried in a discussion comment.

Also, using a modern build system like Vite will often lead to
deprecation warnings due to the older color functions used in Active
Admin's SCSS assets.

To address these issues, I've made some small improvements to Active
Admin's docs. Namely:

- Add a vite_rails section to the installation document.
- Mention the SCSS deprecation warnings in the "gotchas" document, with
  a workaround that specifically works with Vite.

As discussed in
<https://github.com/activeadmin/activeadmin/discussions/8492>, it would
be great to solve the deprecation warnings rather than just hiding them,
but that will require some significant refactoring, which can be a
separate PR. In the meantime I wanted to improve the documentation to at
least point users in the right direction.