Remove `frozen_string_literal` comment from `.arb`
The suggestion to add the `# frozen_string_literal: true` comment at
the top of `.arb` files was a false positive for the following reasons:

- It produced a warning.
- It did not have any effect because `.arb` files are evaluated in a
  different context.

To make this work correctly, the comment should be added at the
framework level instead.

This change eliminates the warning when running specs with
`RUBYOPT="-w"`.

Closes #8597

References:
- rails/rails#43725
- rubocop/rubocop#13699