Enhance main navbar to improve bouncing on macOS
As specified in the Flowbite documentation, the sticky navbar should
utilize the `fixed` class instead of the `sticky` class.

This was causing a poor bouncing effect on macOS on all browser except
Safari.

This commit:
- Replaces the `sticky` class with the `fixed` class, to improve
  bouncing effect
- Adds a top padding to the main body element, to take into
  consideration the space used by the fixed navbar

NOTE: Flowbite documentation suggests to use the `aside` tag instead
for the sidebar, but this breaks the see-through effect, so `div` is
being preserved

Close #8726

Ref:
- https://flowbite.com/docs/components/sidebar/#sidebar-with-navbar