Remove double bottom border on last table row
For when using the table_for component directly its best to remove the bottom border on the last table row for a consistent display. This way a simple border on the table can be added as needed without causing a double border at the bottom. Since the paginated-collection-pagination is optional we have to apply the top border just like we do for the paginated-collection-footer which follows it.