Remove Tabs component (#8762)
The intention was to always remove this component as part of v4. The last component that is not directly used by ActiveAdmin is Panel which will remain as is with no other enhancements. In another major release we may remove Panel. The objective is for users to own their own components, whether directly or from another third party library. This also frees up the global builder method namespace in Arbre for tabs without conflict.