Fix file input styles
Since these were taken from Flowbite to start, there are several issues. The ::file-selector-button was offset and wasn't rendered correctly. The offset was to add spacing between the button and generated browser text that we can't style. It would be preferred as a default to not offset. This applies same styles as other input controls to keep it simple and consistent. The :focus style was removed since it was invalid anyway and not applied by the browser.

This was partially addressed in #8574 but not targeting the right element. This approach is preferred to simplify the styles as the fix in #8574 (display block and overflow hidden) are unnecessary and cause undesired effects e.g. no border radius visible on left side of button.