Simplify checkbox and radio styles
Originally, this required different background sizes due to the checkbox SVG but all 3 SVGs now account for being within a padded square box so we can now set a single background size of 100% for both elements. The checkbox SVGs had to be recreated for that. This also removes a duplicate rule (.dark [type='radio']:checked) because of CSS specificity which we no longer need.